Asymptotic p-value of Exchangeability Using Distance Data

Generates an asymptotic p-value.

distDataLargeP(dist_list)

Arguments

dist_list
The list (length \(B\)) of pairwise distance data. Each element in list should be either a distance matrix or a table recording pairwise distances.

Value

The asymptotic p-value obtained from the weighted convolution of chi-squares distribution.

Details

Generates a weighted convolution of chi-squares distribution of \(V\) statistic by storing the provided list of distance data as an \({N\choose2} \times B\) array, and then using large-\(P\) theory to generate the asymptotic null distribution against which the p-value of observed \(V\) statistic is computed.

Each element of dist_list should be a \(N\times N\) distance matrix.

Dependencies: buildReverse, getChi2Weights, weightedChi2P