A Non-parametric Test for Exchangeability and Homogeneity (Distance List Version)

Computes the p-value of a multivariate dataset, which informs the user if the sample is exchangeable at a given significance level, while simultaneously accounting for feature dependencies. See Aw, Spence and Song (2021) for details.

distDataPValue(dist_list, largeP = FALSE, nruns = 1000)

Arguments

dist_list
The list of distances

largeP
Boolean indicating whether to use large \(P\) asymptotics. Default is FALSE.

nruns
Resampling number for exact test. Default is 1000.

Value

The p-value to be used to test the null hypothesis of exchangeability

Details

This version takes in a list of distance matrices recording pairwise distances between individuals across \(B\) independent features.

Dependencies: distDataLargeP and distDataPermute from auxiliary.R

Examples

# Example
suppressWarnings(require(doParallel))

#> Loading required package: doParallel

#> Loading required package: foreach

#> Loading required package: iterators

#> Loading required package: parallel

registerDoParallel(cores = 2)

N <- 100
num_rows <- choose(N,2)
B <- 20
dist_list <- list()
for (b in 1:B) {
  # generate fake data 
  fake_samples <- replicate(10, rbinom(N,1,0.5))
 
  # compute pairwise distance 
  fake_dist <- as.matrix(dist(fake_samples, method = "manhattan"))
 
  # add to list
  dist_list[[b]] <- fake_dist
}

distDataPValue(dist_list)

#> [1] 0.503

# should be larger than 0.05

stopImplicitCluster()